refactor : Fix misspellings and typos
- Update language files
- Add new keys
