+++ src/locales/en.json
@@ -53,7 +53,8 @@
     "btn_start_over": "Start over",
     "btn_mint_again": "Mint Again",
     "btn_swap": "Swap",
-    "btn_approve": "Approve"
+    "btn_approve": "Approve",
+    "btn_close": "Close"
   },
   "errors": {
     "install_wallet": "No ethereum provider found, please install metamask or similar",
@@ -72,7 +73,7 @@
     "add_video": "Add a Video",
     "add_link": "Add a main link",
     "loading": "Loading...",
-    "new_post_plh": "What's happening?",
+    "new_post_plh": "Now is a great time to write something beautiful...",
     "likes": "Likes",
     "dislikes": "Dislikes",
     "comments": "Comments",
@@ -89,11 +90,11 @@
     "cost": "Cost",
     "minting": "Minting...",
     "minted": "Minted",
-    "minted_success": "You have successfully minted an NTF",
+    "minted_success": "You have successfully minted an NFT",
     "approving": "Approving...",
     "approved": "Approved",
     "block_explorer": "Block Explorer",
-    "check_ntf_on": "Check out your NFT on",
+    "check_nft_on": "Check out your NFT on",
     "its_magic": "✨It's magic✨",
     "balance": "Balance",
     "enter_amount": "Enter an amount",
@@ -108,7 +109,9 @@
     "quickmintlabel": "Mint professionally, with more options, learning it, and step-by-step.",
     "getmcaw": "Get mCAW",
     "getmcawlarge": "Get MintableCAW",
-    "testnet_msg": "Connect to Goerli testnet to test the process"
+    "testnet_msg": "Connect to Goerli testnet to test the process",
+    "buycaw": "Buy / Swap CAW",
+    "linksmenu": "Quick Links"
   },
   "common": {
     "amountgtzero": "Amount must be greater than zero"
@@ -128,8 +131,9 @@
     "decentralized": "Decentralized",
     "social": "Social",
     "caw_is": "CAW IS",
-    "btn_mint": "Mint your",
-    "btn_mint_tooltip": "Minting an NTF-Username is the new way to enter to a revolutionary world. Join us!",
+    "btn_mint": "Mint Username",
+    "btn_register": "Start here",
+    "btn_mint_tooltip": "Minting an NFT-Username is the new way to enter to a revolutionary world. Join us!",
     "censorship_title": "A censorship-resistant, decentralized messaging protocol",
     "blockchain_title": "Blockchain Technology",
     "blockchain_label": "Blockchain is a shared, immutable ledger that facilitates the process of recording transactions and tracking assets in a business network. It helping to ensure that no single entity controls the database.",
@@ -177,10 +181,10 @@
     "message": "Mint your username and join the revølution",
     "step_title": "First step is to connect your",
     "step_title_wallet": "wallet",
-    "step_description": "Minting an NTF username is the new way to enter to a revolutionary world.",
-    "ntf_state_q": "What is an NTF?",
-    "ntf_state_a": "NTF refers to a Non-Fungible Token and is a digital asset that is unique and stored on a blockchain that allows for the secure transfer of ownership of digital assets.",
-    "ntf_state_ai": "Ownership of an NTF username will provide access to the associated account.",
+    "step_description": "Minting an NFT username is the new way to enter to a revolutionary world.",
+    "nft_state_q": "What is an NFT?",
+    "nft_state_a": "NFT refers to a Non-Fungible Token and is a digital asset that is unique and stored on a blockchain that allows for the secure transfer of ownership of digital assets.",
+    "nft_state_ai": "Ownership of an NFT username will provide access to the associated account.",
     "already_minted": "Already have your CAW-NFT?",
     "username_already_taken": "already taken",
     "username_invalid": "username must be : alphanumeric, lowercase, no spaces, non special characters",
@@ -195,11 +199,11 @@
     "label_cost_willbe": "it will be.",
     "wizard_step1_lb1": "All user activity, social and financial flows through your NFT username",
     "wizard_step1_lb2": "A wallet can own several usernames",
-    "wizard_step1_lb3": "You can mint as many as you can afford it",
+    "wizard_step1_lb3": "You can mint as many as you can afford",
     "upding_balance": "Updating balance...",
     "main_balance": "Main balance",
     "caw_balance_req_lb": "You need to have enough mCAW to mint a username. You can mint mCAW here : ",
-    "enter_ntf_input_lb": "Enter your new NTF-Username",
+    "enter_nft_input_lb": "Enter your new NFT-Username",
     "user_acceptance_title": "Before continuing, you should know and accept the following",
     "user_acceptance_mchk_1": "This is a",
     "user_acceptance_mchk_dapp": "decentralized application",
@@ -221,5 +225,33 @@
     "get_goerli": "Retrieve some Goerli ETH from a faucet, i.g. : ",
     "approve_req_msg": "Once you mint, you will need to approve the transaction to transfer your CAW tokens,otherwise your funds will be locked in the contract.",
     "alert_cf_mint": "Do you want to mint {$0} mCAW?"
+  },
+  "calc": {
+    "title": "Cost and burns at different MC's",
+    "action": "ACTION",
+    "burns_dist": "BURN / DISTRIBUTION",
+    "MCTD": "Market Cap as of today",
+    "MC50M": "Market Cap at $50 million",
+    "MC1B": "Market Cap at $1 billion",
+    "MC10B": "Market Cap at $10 billion",
+    "tbl_caption": "Forecast capitalization in USD",
+    "mint_action": "Mint {0}{1} Character{2} username",
+    "follow_acc": "Follow Account",
+    "follow_acc_burn": "paid 80/20% to account and stakepool",
+    "send_caw": "Send CAW",
+    "send_caw_burn": "paid 100% to stakepool",
+    "like_caw": "Like CAW",
+    "like_caw_burn": "paid 80/20% to account and stakepool",
+    "recaw_burn": "paid 50/50% to account and stakepool"
+  },
+  "menu": {
+    "menu": "Menu",
+    "getmCaw": "Get mCAW",
+    "join_tg": "Join us on Telegram",
+    "contribution": "Want to contribute?",
+    "bugs_report": "Bugs and feature requests",
+    "firsttweb3": "First time using Web3?",
+    "learn_more": "Learn more",
+    "exp_exchanges": "Explore Exchanges"
   }
 }
