+++ src/locales/es.json
@@ -44,7 +44,8 @@
     "btn_start_over": "Comenzar de nuevo",
     "btn_mint_again": "Acuñar otro",
     "btn_swap": "Intercambiar",
-    "btn_approve": "Aprobar"
+    "btn_approve": "Aprobar",
+    "btn_close": "Cerrar"
   },
   "errors": {
     "install_wallet": "No se encontró el proveedor. Por favor, instale metamask o su monedero similar.",
@@ -63,7 +64,7 @@
     "add_video": "Añadir video",
     "add_link": "Añadir enlace principal",
     "loading": "Cargando...",
-    "new_post_plh": "¿Qué estás pensando?",
+    "new_post_plh": "Es un buen día para compartir algo bonito...",
     "likes": "Me gusta",
     "dislikes": "No me gusta",
     "comments": "Comentarios",
@@ -76,7 +77,7 @@
     "layout": "Diseño",
     "fullscreen": "Pantalla completa",
     "exit_fullscreen": "Salir de pantalla completa",
-    "add_long_caw": "CAW largo",
+    "add_long_caw": "Artículo CAW",
     "cost": "Costo",
     "minting": "Acuñando...",
     "minted": "Acuñado",
@@ -84,7 +85,7 @@
     "approving": "Aprobando...",
     "approved": "Aprobado",
     "block_explorer": "Explorador de bloques",
-    "check_ntf_on": "Compruebe tu NFT en",
+    "check_nft_on": "Compruebe tu NFT en",
     "its_magic": "✨Es fantástico✨",
     "balance": "Balance",
     "enter_amount": "Introduzca la cantidad",
@@ -99,7 +100,9 @@
     "quickmintlabel": "Acuñar de manera profesional, con más opciones, aprendiendo y paso a paso.",
     "getmcaw": "Obtener mCAW",
     "getmcawlarge": "Obtener MintableCAW",
-    "testnet_msg": "Conectarse a la testnet de Goerli para continuar con el proceso"
+    "testnet_msg": "Conectarse a la testnet de Goerli para continuar con el proceso",
+    "buycaw": "Comprar / Intercambiar CAW",
+    "linksmenu": "Enlaces de interés"
   },
   "common": {
     "amountgtzero": "La cantidad debe ser mayor que cero"
@@ -120,7 +123,8 @@
     "social": "Social",
     "caw_is": "CAW ES",
     "btn_mint": "Crea el tuyo",
-    "btn_mint_tooltip": "Acuñar un NTF es la nueva forma de entrar al nuevo y revolucionario mundo. ¡Únase a nosotros!",
+    "btn_register": "Comenzar",
+    "btn_mint_tooltip": "Acuñar un NFT es la nueva forma de entrar al nuevo y revolucionario mundo. ¡Únase a nosotros!",
     "censorship_title": "Un protocolo de mensajería descentralizado y resistente a la censura",
     "blockchain_title": "Tecnología Blockchain",
     "blockchain_label": "Blockchain es un libro contable compartido e inmutable que facilita el proceso de registro de transacciones y seguimiento de activos en una red de negocios. Ayuda a garantizar que ninguna entidad controle los datos.",
@@ -168,10 +172,10 @@
     "message": "Acuñar tu NFT y únete a la revølución.",
     "step_title": "El primer paso es conectar tu",
     "step_title_wallet": "monedero",
-    "step_description": "Acuñar un usuario-NTF es la nueva forma de entrar al nuevo y revolucionario mundo.",
-    "ntf_state_q": "¿Qué es un NTF?",
-    "ntf_state_a": "NTF signitica Token No Fungible y es un activo digital que es único y se almacena en una blockchain que permite la transferencia segura de la propiedad de activos digitales.",
-    "ntf_state_ai": "Poseer un nombre de usuario NTF dará acceso a la cuenta asociada.",
+    "step_description": "Acuñar un usuario-NFT es la nueva forma de entrar al nuevo y revolucionario mundo.",
+    "nft_state_q": "¿Qué es un NFT?",
+    "nft_state_a": "NFT signitica Token No Fungible y es un activo digital que es único y se almacena en una blockchain que permite la transferencia segura de la propiedad de activos digitales.",
+    "nft_state_ai": "Poseer un nombre de usuario NFT dará acceso a la cuenta asociada.",
     "already_minted": "¿Ya tienes un CAW-NFT?",
     "username_already_taken": "ya está tomado",
     "username_invalid": "Nombre de usuario debe ser : alfanumérico, minúsculas, sin espacios, sin caracteres especiales",
@@ -190,7 +194,7 @@
     "upding_balance": "Actualizando balance...",
     "main_balance": "Balance principal",
     "caw_balance_req_lb": "Debe tener suficiente mCAW para acuñar un nombre de usuario. Puedes acuñar por mCAW aquí :",
-    "enter_ntf_input_lb": "Ingresa tu nuevo nombre de usuario NTF",
+    "enter_nft_input_lb": "Ingresa tu nuevo nombre de usuario NFT",
     "user_acceptance_title": "Antes de continuar, debe conocer y aceptar lo siguiente",
     "user_acceptance_mchk_1": "Esta es una",
     "user_acceptance_mchk_dapp": "aplicación descentralizada",
@@ -211,5 +215,33 @@
     "get_goerli": "Solicitar Goerli ETH de una faucet, por ejemplo : ",
     "approve_req_msg": "Una vez que acuñado, debe que aprobar la transacción para transferir sus tokens CAW, de lo contrario sus fondos quedarán bloqueados en el contrato.",
     "alert_cf_mint": "¿Desea acuñar {$0} mCAW?"
+  },
+  "calc": {
+    "title": "Costo y quemas en diferente Capitalización de Mercado",
+    "action": "ACCIÓN",
+    "burns_dist": "Quemas y distribución",
+    "MCTD": "Capitalización a la fecha",
+    "MC50M": "Capitalización a $50 millones",
+    "MC1B": "Capitalización a $1 billón",
+    "MC10B": "Capitalización a $10 mil millones",
+    "tbl_caption": "Previsiones de capitalización en USD",
+    "mint_action": "Acuñar usuario de {0}{1} letra{2}",
+    "follow_acc": "Seguir una cuenta",
+    "follow_acc_burn": "Pago 80/20 % a la cuenta y stakepool",
+    "send_caw": "Crear CAW",
+    "send_caw_burn": "Pago 100% a stakepool",
+    "like_caw": "Like CAW",
+    "like_caw_burn": "Pago 80/20% a la cuenta y stakepool",
+    "recaw_burn": "Pago 50/50 a la cuenta y stakepool"
+  },
+  "menu": {
+    "menu": "Menu",
+    "getmCaw": "Obtener mCAW",
+    "join_tg": "Únetenos en Telegram",
+    "contribution": "Desea contribuir?",
+    "bugs_report": "Fallos o nuevas funciones",
+    "firsttweb3": "Nuevo en Web3?",
+    "learn_more": "Aprende más",
+    "exp_exchanges": "Ver en intercambios"
   }
 }
